Session 7
Groups are nearly done. The user is now presented a list of groups upon opening the website. This list only has the _id and name of the group. When the user selects a group, all the data of that group is loaded. This should save on data and improve scalability. Perhaps I could implement a group-caching system. Adding and removing cards is not possible anymore due to the changes in organization. This will be re-added later.