Session 10
Added user system. Users can now login with Google. Groups are not public anymore. Instead, users need to enter a code to view a group. Users can create and join groups. All of this is linked to the users account. Polished UI a bit, and refusing to refactor backend for functions I'm not using anumore.