Session 11
Cleaned up a bit on the front-end. The biggest update was adding a 'group information' panel. It shows the group's name, the description, the join code, the date-of-creation and a list of participants. Users can now leave the group from this panel. The code is disgusting though, don't look at it too closely.