Session 12
Made some final tweaks to the front end. Back end is still a disgusting (but functional) mess. I added a feature that let users click on a card to see an expanded view of it. I've been postponing this for almost the entire project now, but with my modal system, this was fairly rudementary. I would've wanted the expanded card to have the same colours as the card, but that would require me to restructure the database (again) and write a whole bunch of code for a minor improvement. I'd say that the project is pretty much finished now. I might add some final tweaks later, maybe some styling or bugfixes, but this is approximately what the final product looks like. I've made many mistakes, but I learned so much. I'm pretty proud of this project, even if it's a mess,
looking back.